+++ src/app/(routes)/blog/[slug]/_components/contentful-image.tsx
