+++ src/app/(routes)/blog/_components/article-snippet-card.tsx
@@ -1,10 +1,9 @@
 import { BlogPostType } from "@/app/_lib/contentful";
-import ContentfulImage from "./[slug]/ContentfulImage";
-import { Asset } from "contentful";
-import { MetaInfo } from "./[slug]/MetaInfo";
 import { Text } from "@/app/_components/text";
 import { documentToPlainTextString } from "@contentful/rich-text-plain-text-renderer";
 import Link from "next/link";
+import ContentfulImage from "../[slug]/_components/contentful-image";
+import { MetaInfo } from "../[slug]/_components/meta-info";
 
 export default function ArticleSnippetCard({
   article,
